Added test classes. Refresh code.

+++ src/main/java/org/example/service/MoneyService.java
@@ -2,13 +2,17 @@ package org.example.service;
 
 
 import lombok.RequiredArgsConstructor;
+import org.example.controller.response.MoneyResponse;
 import org.example.entity.Money;
+import org.example.exception.BalanceException;
+import org.example.exception.EntityNotExistException;
+import org.example.exception.ErrorCode;
 import org.example.repository.MoneyRepository;
 import org.example.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.example.utils.Messages;
 import org.springframework.stereotype.Service;
-import org.example.controller.response.MoneyResponse;
 
+import javax.transaction.Transactional;
 import java.sql.Timestamp;
 import java.util.Date;
 
@@ -21,32 +25,38 @@ public class MoneyService {
     private final UserRepository userRepository;
 
     public MoneyResponse findByUserId(Long userId) {
-        if (moneyRepository.findMoneyByUserId(userId) == null) return null;
-        else return new MoneyResponse(moneyRepository.findMoneyByUserId(userId));
+        Money money = moneyRepository.findMoneyByUserId(userId);
+        return (money == null) ? null : new MoneyResponse(money);
     }
 
-    public String deleteMoney(Long userId) {
-        moneyRepository.deleteById(userId);
-        return "Wallet deleted.";
+    public void deleteMoney(Long userId) {
+        Money moneyByUserId = moneyRepository.findMoneyByUserId(userId);
+        if (moneyByUserId != null) {
+            moneyRepository.deleteById(userId);
+        } else {
+            throw new EntityNotExistException(String.format(Messages.WALLET_NOT_EXIST, userId));
+        }
     }
 
     public MoneyResponse createMoney(Long userId) {
-        Money money = new Money(userRepository.findUserById(userId), 0f, 0f);
+        Money money = new Money(userRepository.findUserById(userId));
         money.setLastChangeDate(new Timestamp(new Date().getTime()));
         return new MoneyResponse(moneyRepository.save(money));
     }
 
+    @Transactional
     public float update(Long userId, Float transfer) {
         Money existingMoney = moneyRepository.findMoneyByUserId(userId);
         long actualVersion = existingMoney.getVersion();
         float newBalance = existingMoney.getBalance() + transfer;
-        if (newBalance < 0) throw new RuntimeException("Balance < 0");
+        if (newBalance < 0) throw new BalanceException(ErrorCode.BALANCE_LESS_0.toString());
         int result = moneyRepository.updateById(userId, newBalance, actualVersion, transfer, new Timestamp(new Date().getTime()));
-        if (result == 0) throw new RuntimeException("Not updated, try again");
+        if (result == 0) throw new BalanceException(ErrorCode.BALANCE_NOT_UPDATED.toString());
         return newBalance;
     }
+
     public MoneyResponse updateMoney(Long userId, Float transfer) {
-        update(userId,transfer);
+        update(userId, transfer);
         return new MoneyResponse(moneyRepository.findMoneyByUserId(userId));
     }
 }
